Update PromptInjection.md
Opened this as a separate issues because it could lead to some debate... I recommend deleting examples of vulnerability (or rewording). As worded they do not even involve the LLM. As described prompt injection involves only text going into the LLM while these examples talk about the plugin ingesting and processing the text that leads to the plugin's exploit directly.